+++ tests/PerFiFeatureTest/Domain/TransactionContext.php
@@ -12,10 +12,10 @@ use PerFi\Domain\Transaction\CommandHandler\ExecutePayment;
 use PerFi\Domain\Transaction\CommandHandler\ExecuteRefund;
 use PerFi\Domain\Transaction\Command\Pay;
 use PerFi\Domain\Transaction\Command\Refund;
-use PerFi\Domain\Transaction\EventSubscriber\CreditAssetAccountWhenPaymentMade;
-use PerFi\Domain\Transaction\EventSubscriber\CreditExpenseAccountWhenTransactionRefunded;
-use PerFi\Domain\Transaction\EventSubscriber\DebitAssetAccountWhenTransactionRefunded;
-use PerFi\Domain\Transaction\EventSubscriber\DebitExpenseAccountWhenPaymentMade;
+use PerFi\Domain\Account\EventSubscriber\CreditAssetAccountWhenPaymentMade;
+use PerFi\Domain\Account\EventSubscriber\CreditExpenseAccountWhenTransactionRefunded;
+use PerFi\Domain\Account\EventSubscriber\DebitAssetAccountWhenTransactionRefunded;
+use PerFi\Domain\Account\EventSubscriber\DebitExpenseAccountWhenPaymentMade;
 use PerFi\Domain\Transaction\EventSubscriber\MarkRefundedTransactionAsRefundedWhenTransactionRefunded;
 use PerFi\Domain\Transaction\Event\PaymentMade;
 use PerFi\Domain\Transaction\Event\TransactionRefunded;
