+++ tests/PerFiUnitTest/Domain/Account/EventSubscriber/DebitAssetAccountWhenTransactionRefundedTest.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace PerFiUnitTest\Domain\Transaction\EventSubscriber;
+namespace PerFiUnitTest\Domain\Account\EventSubscriber;
 
 use Mockery as m;
 use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
@@ -9,7 +9,7 @@ use PHPUnit\Framework\TestCase;
 use PerFiUnitTest\Traits\EventBusTrait;
 use PerFiUnitTest\Traits\TransactionTrait;
 use PerFi\Domain\Account\Event\DestinationAccountDebited;
-use PerFi\Domain\Transaction\EventSubscriber\DebitAssetAccountWhenTransactionRefunded;
+use PerFi\Domain\Account\EventSubscriber\DebitAssetAccountWhenTransactionRefunded;
 use PerFi\Domain\Transaction\Event\TransactionRefunded;
 use SimpleBus\Message\Bus\MessageBus;
